fix: js파일 수정 2024-09-14 빌드 포매팅 및 csrf&cors 세팅 위치 이동
